+++ app/src/main/java/frascog/smartmirror/Weather/Datum_.java
@@ -16,10 +16,10 @@ public class Datum_ {
     private String icon;
     @SerializedName("precipIntensity")
     @Expose
-    private int precipIntensity;
+    private double precipIntensity;
     @SerializedName("precipProbability")
     @Expose
-    private int precipProbability;
+    private double precipProbability;
     @SerializedName("temperature")
     @Expose
     private double temperature;
@@ -168,7 +168,7 @@ public class Datum_ {
      * @return
      *     The precipIntensity
      */
-    public int getPrecipIntensity() {
+    public double getPrecipIntensity() {
         return precipIntensity;
     }
 
@@ -177,7 +177,7 @@ public class Datum_ {
      * @param precipIntensity
      *     The precipIntensity
      */
-    public void setPrecipIntensity(int precipIntensity) {
+    public void setPrecipIntensity(double precipIntensity) {
         this.precipIntensity = precipIntensity;
     }
 
@@ -191,7 +191,7 @@ public class Datum_ {
      * @return
      *     The precipProbability
      */
-    public int getPrecipProbability() {
+    public double getPrecipProbability() {
         return precipProbability;
     }
 
@@ -200,7 +200,7 @@ public class Datum_ {
      * @param precipProbability
      *     The precipProbability
      */
-    public void setPrecipProbability(int precipProbability) {
+    public void setPrecipProbability(double precipProbability) {
         this.precipProbability = precipProbability;
     }
